fix: no circular reference anymore
