+++ src/components/AppComponent.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import * as classNames from 'classnames'
 import { isMobile } from '../browser.js'
-import { store } from '../store.js'
+import store from '../store.js'
 import globals from '../globals.js'
 import { handleGesture } from '../shortcuts.js'
