+++ src/components/Child.js
@@ -4,7 +4,7 @@ import * as classNames from 'classnames'
 import { DragSource, DropTarget } from 'react-dnd'
 import { getEmptyImage } from 'react-dnd-html5-backend'
 import { isMobile } from '../browser.js'
-import { store } from '../store.js'
+import store from '../store.js'
 import globals from '../globals.js'
 
 // components
