+++ src/components/Children.js
@@ -4,7 +4,7 @@ import * as classNames from 'classnames'
 import assert from 'assert'
 import * as evaluate from 'static-eval'
 import { DropTarget } from 'react-dnd'
-import { store } from '../store.js'
+import store from '../store.js'
 import { isMobile } from '../browser.js'
 import { shortcutById } from '../shortcuts.js'
 import globals from '../globals.js'
