+++ src/components/Link.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import { store } from '../store.js'
+import store from '../store.js'
 
 // util
 import {
