+++ src/components/Search.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import * as debounce from 'lodash.debounce'
-import { store } from '../store.js'
+import store from '../store.js'
 import { isMobile } from '../browser.js'
 import { shortcutById } from '../shortcuts.js'
 import * as assert from 'assert'
