+++ src/components/SearchChildren.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import { store } from '../store.js'
+import store from '../store.js'
 
 // components
 import { Children } from './Children.js'
