+++ src/components/Superscript.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import { store } from '../store.js'
+import store from '../store.js'
 import { isMobile } from '../browser.js'
 
 // components
