+++ src/util.js
@@ -2,7 +2,8 @@
 
 import * as htmlparser from 'htmlparser2'
 import { clientId, isMac, isMobile } from './browser.js'
-import { fetch, store, sync, syncRemote } from './store.js'
+import { fetch, sync, syncRemote } from './store-utils.js'
+import store from './store.js'
 import globals from './globals.js'
 import { handleKeyboard } from './shortcuts.js'
